aws_cdk_static_website/requirements.txt
-e .
aws.cdk.aws_s3
aws.cdk.aws_s3_deployment
aws.cdk.aws_cloudfront